+++ Documentation/SwiftUI MVVM Architecture/Example Project/SwiftUIMVVMExample/Gateways/Posts Gateway/PostsEntity.swift
@@ -16,10 +16,10 @@ struct PostsEntity {
     // MARK: Post
     @MemberwiseCodable
     struct Post: Hashable, Identifiable, Decodable {
-        @MWCKey("id") let id: Int
-        @MWCKey("userId") let userID: Int
-        @MWCKey("title") let title: String
-        @MWCKey("body") let body: String
+        @MWCCodingKey("id") let id: Int
+        @MWCCodingKey("userId") let userID: Int
+        @MWCCodingKey("title") let title: String
+        @MWCCodingKey("body") let body: String
     }
 
     // MARK: Mock