+++ Tests/VCoreTests/Tests/Services and Managers/Multi Part Form Data Builder/MultiPartFormDataBuilderTests.swift
@@ -15,7 +15,7 @@ final class MultipartFormDataBuilderTests: XCTestCase {
 
     @MemberwiseCodable
     private struct JSONPart: Encodable {
-        @MWCKey("key") let key: String
+        @MWCCodingKey("key") let key: String
     }
 
     // MARK: Tests